+++ nAR/nAR.c
@@ -9,9 +9,8 @@
 int main(int argc, char **argv) {
     // Gets arguments from command line
     int i, j[MAX_PARAMS_COUNT];
-    for(i = 0; i < MAX_PARAMS_COUNT; i++) {
+    for(i = 0; i < MAX_PARAMS_COUNT; i++)
         j[i] = i < argc - 2 ? atoi(argv[i + 2]) : 0;
-    }
     // Loads script and gets necessary count of registers and qubits
     qscript r = load_script(argv[1], false);
     // Allocates driver