+++ nAR/nya/operations.h
@@ -4,6 +4,13 @@
 #define MAX_TASK_LENGTH 1<<2
 #define MAX_PARAMS_COUNT 1<<2
 
+typedef enum _nyanTag {
+    TAG_NONE        = 0,
+    TAG_ARGUMENT    = 1<<0,
+    TAG_RESERVED    = 1<<1,
+    TAG_REFERENCE   = 1<<2
+} nyanTag;
+
 typedef enum _nyanParam {
     TYPE_VOID,
     TYPE_VAL,
@@ -18,26 +25,26 @@ typedef struct _nyanTask {
 } nyanTask;
 
 typedef struct _nyanOperation {
-    int     code;
-    int     values[MAX_PARAMS_COUNT];
-    bool    specials[MAX_PARAMS_COUNT];
+    int         code;
+    int         values[MAX_PARAMS_COUNT];
+    unsigned    specials[MAX_PARAMS_COUNT];
 } nyanOperation;
 
 static const nyanTask nyanTasks[] = {
-    { "end",    { TYPE_BIT } },
-    { "mov",    { TYPE_BIT, TYPE_BIT } },
-    { "put",    { TYPE_BIT, TYPE_VAL } },
-    { "cmp",    { TYPE_BIT } },
+    { "end",    { TYPE_VAL } },
+    { "mov",    { TYPE_BIT, TYPE_VAL } },
+    { "cmp",    { TYPE_BIT, TYPE_VAL } },
+    { "jmp",    { TYPE_LABEL } },
     { "je",     { TYPE_LABEL } },
     { "jne",    { TYPE_LABEL } },
     { "jg",     { TYPE_LABEL } },
     { "jge",    { TYPE_LABEL } },
     { "jl",     { TYPE_LABEL } },
     { "jle",    { TYPE_LABEL } },
-    { "add",    { TYPE_BIT, TYPE_BIT } },
-    { "sub",    { TYPE_BIT, TYPE_BIT } },
-    { "mul",    { TYPE_BIT, TYPE_BIT } },
-    { "div",    { TYPE_BIT, TYPE_BIT } },
+    { "add",    { TYPE_BIT, TYPE_VAL } },
+    { "sub",    { TYPE_BIT, TYPE_VAL } },
+    { "mul",    { TYPE_BIT, TYPE_VAL } },
+    { "div",    { TYPE_BIT, TYPE_VAL } },
     { "m",      { TYPE_QUBIT } },
     { "h",      { TYPE_QUBIT } },
     { "x",      { TYPE_QUBIT } },
@@ -48,8 +55,8 @@ static const nyanTask nyanTasks[] = {
 static const enum _nyanOperations {
     OP_END,
     OP_MOV,
-    OP_PUT,
     OP_CMP,
+    OP_JMP,
     OP_JE,
     OP_JNE,
     OP_JG,