+++ nAR/CHANGELOG.md
@@ -2,16 +2,25 @@
   <img src="https://raw.githubusercontent.com/thzoid/vscode-nyancat/master/images/icon_small.png">
 </p>
 
-# nAR Changelog
+# nAR Change Log
 
-## 1.1
+## 2.0
+
+### Features
+ - Full support for nyancat He;
+ - Increased the size of a classical register from 1 byte to 4 bytes.
 
 ### Bug fixes
- - Increased the max size of numeric values on `.nya` scripts to match the documentation.
+ - Fixed some bad-addressed error messages.
+
+## 1.1
 
 ### Features
  - Hello World code;
  - Function to run simple algorithms (no arguments).
+
+### Bug fixes
+ - Increased the max size of numeric values on `.nya` scripts to match the documentation.
 
 ## 1.0